+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
+## [1.1.0]
+- Add Flow schema
+- Add system swap helper
+- Add http client get-responses and clear-requests
+
 ## [1.0.0]
+- Remove responses from function pararemeters
+
+## [0.1.0]
 - Moved core code from nu-algebraic-data-types
 - Moved helpers from purgatory code
-- Remove responses from function pararemeters